+++ src/steps/resource-manager/batch/index.ts
@@ -19,9 +19,13 @@ import {
   BatchAccountRelationships,
   STEP_RM_BATCH_ACCOUNT,
   STEP_RM_BATCH_POOL,
-  // STEP_RM_BATCH_APPLICATION
+  STEP_RM_BATCH_APPLICATION,
 } from './constants';
-import { createBatchAccountEntity, createBatchPoolEntity } from './converters';
+import {
+  createBatchAccountEntity,
+  createBatchApplicationEntity,
+  createBatchPoolEntity,
+} from './converters';
 import { resourceGroupName } from '../../../azure/utils';
 
 export * from './constants';
@@ -100,6 +104,49 @@ export async function fetchBatchPools(
   );
 }
 
+export async function fetchBatchApplications(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const accountEntity = await jobState.getData<Entity>(ACCOUNT_ENTITY_TYPE);
+
+  const webLinker = createAzureWebLinker(accountEntity.defaultDomain as string);
+  const client = new BatchClient(instance.config, logger);
+
+  await jobState.iterateEntities(
+    { _type: BatchEntities.BATCH_ACCOUNT._type },
+    async (batchAccountEntity) => {
+      const { id, name } = batchAccountEntity;
+      const resourceGroup = resourceGroupName(id, true)!;
+
+      await client.iterateBatchApplications(
+        ({
+          resourceGroupName: resourceGroup,
+          batchAccountName: name,
+        } as unknown) as {
+          resourceGroupName: string;
+          batchAccountName: string;
+        },
+        async (batchApplication) => {
+          const batchApplicationEntity = createBatchApplicationEntity(
+            webLinker,
+            batchApplication,
+          );
+          await jobState.addEntity(batchApplicationEntity);
+
+          await jobState.addRelationship(
+            createDirectRelationship({
+              _class: RelationshipClass.HAS,
+              from: batchAccountEntity,
+              to: batchApplicationEntity,
+            }),
+          );
+        },
+      );
+    },
+  );
+}
+
 export const batchSteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
@@ -122,5 +169,19 @@ export const batchSteps: Step<
       STEP_RM_BATCH_ACCOUNT,
     ],
     executionHandler: fetchBatchPools,
+  },
+  {
+    id: STEP_RM_BATCH_APPLICATION,
+    name: 'Batch Applications',
+    entities: [BatchEntities.BATCH_APPLICATION],
+    relationships: [
+      BatchAccountRelationships.BATCH_ACCOUNT_HAS_BATCH_APPLICATION,
+    ],
+    dependsOn: [
+      STEP_AD_ACCOUNT,
+      STEP_RM_RESOURCES_RESOURCE_GROUPS,
+      STEP_RM_BATCH_ACCOUNT,
+    ],
+    executionHandler: fetchBatchApplications,
   },
 ];