+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 - Added `azure_resource_group|has|azure_batch_account` relationships
 - Added `azure_batch_pool` entities
 - Added `azure_batch_account|has|azure_batch_pool` relationships
+- Added `azure_batch_application` entities
+- Added `azure_batch_account|has|azure_batch_application` relationships
 
 ## 5.3.1 - 2020-09-21